webapp/requirements.txt
alembic==1.7.7
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
dominate==2.6.0
Flask==2.1.2
Flask-Bootstrap4==4.0.2
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
marshmallow==3.15.0
marshmallow-sqlalchemy==0.28.0
packaging==21.3
psycopg2-binary==2.8.6
pyparsing==3.0.8
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.36
urllib3==1.26.9
visitor==0.1.3
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
